Adding helper for badges
