+++ plugins/pipelines/badge.js
@@ -4,40 +4,8 @@ const joi = require('joi');
 const schema = require('screwdriver-data-schema');
 const workflowParser = require('screwdriver-workflow-parser');
 const idSchema = schema.models.pipeline.base.extract('id');
-const { makeBadge } = require('badge-maker');
+const { getPipelineBadge } = require('./helper');
 
-/**
- * Generate Badge Format
- * @method getLabels
- * @param  {Object} statusColor             Mapping for status and color
- * @param  {Array}  [buildsStatus=[]]       An array of builds
- * @param  {String} [label='pipeline']         Subject of the badge
- * @return {Object}
- */
-function getLabels({ statusColor, buildsStatus = [], label = 'pipeline' }) {
-    const counts = {};
-    const parts = [];
-    let worst = 'lightgrey';
-
-    const levels = Object.keys(statusColor);
-
-    buildsStatus.forEach(status => {
-        counts[status] = (counts[status] || 0) + 1;
-    });
-
-    levels.forEach(status => {
-        if (counts[status]) {
-            parts.push(`${counts[status]} ${status}`);
-            worst = statusColor[status];
-        }
-    });
-
-    return {
-        label,
-        message: parts.length > 0 ? parts.join(', ') : 'unknown',
-        color: worst
-    };
-}
 /**
  * DFS the workflowGraph from the start point
  * @method dfs
@@ -93,17 +61,11 @@ module.exports = config => ({
                 statusColor
             };
 
-            const getBadge = badgeObject => {
-                const labels = getLabels(badgeObject);
-
-                return makeBadge(labels);
-            };
-
             return factory
                 .get(request.params.id)
                 .then(pipeline => {
                     if (!pipeline) {
-                        return h.response(getBadge(badgeConfig));
+                        return h.response(getPipelineBadge(badgeConfig));
                     }
 
                     return pipeline.getEvents({ sort: 'ascending' }).then(allEvents => {
@@ -111,7 +73,7 @@ module.exports = config => ({
                             const lastEvent = events.pop();
 
                             if (!lastEvent) {
-                                return h.response(getBadge(badgeConfig));
+                                return h.response(getPipelineBadge(badgeConfig));
                             }
 
                             return lastEvent.getBuilds().then(builds => {
@@ -134,7 +96,7 @@ module.exports = config => ({
                                 }
 
                                 return h.response(
-                                    getBadge(
+                                    getPipelineBadge(
                                         Object.assign(badgeConfig, {
                                             buildsStatus,
                                             label: pipeline.name
@@ -147,7 +109,7 @@ module.exports = config => ({
                         return getLastEffectiveEvent(allEvents);
                     });
                 })
-                .catch(() => h.response(getBadge(badgeConfig)));
+                .catch(() => h.response(getPipelineBadge(badgeConfig)));
         },
         validate: {
             params: joi.object({
